Move PlayerRef from services/gacha.py to services/players.py
PlayerRef is a two-field "player's Telegram identity" dataclass —
nothing about pity, rarity, or pull mechanics is in it. It only lived
in gacha.py because pull_single/pull_ten were its first consumer.
services/players.py's own docstring already states the exact reasoning
that should have put it there from the start.

Part of #58
