+++ src/ley_shards_bot/commands/gacha.py
@@ -22,6 +22,7 @@ from ley_shards_bot.commands.helpers.scoping import NOT_IN_DM_MESSAGE, in_privat
 from ley_shards_bot.db import session_scope
 from ley_shards_bot.models import Rarity
 from ley_shards_bot.services import gacha
+from ley_shards_bot.services.players import PlayerRef
 
 if TYPE_CHECKING:
     from collections.abc import Callable
@@ -95,7 +96,7 @@ async def _announce_rare_pull(
 async def _attempt_pull(
     session: Session,
     message: Message,
-    player: gacha.PlayerRef,
+    player: PlayerRef,
     pull_fn: Callable[..., _PullResultT],
 ) -> _PullResultT | None:
     """Shared `/pull`/`/pull10` flow: resolve the standard banner, attempt
@@ -129,7 +130,7 @@ async def pull_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> No
         await message.reply_text(NOT_IN_DM_MESSAGE)
         return
 
-    player = gacha.PlayerRef(user.id, username=user.username)
+    player = PlayerRef(user.id, username=user.username)
     with session_scope() as session:
         outcome = await _attempt_pull(session, message, player, gacha.pull_single)
         if outcome is None:
@@ -152,7 +153,7 @@ async def pull_ten_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -
         await message.reply_text(NOT_IN_DM_MESSAGE)
         return
 
-    player = gacha.PlayerRef(user.id, username=user.username)
+    player = PlayerRef(user.id, username=user.username)
     with session_scope() as session:
         outcomes = await _attempt_pull(session, message, player, gacha.pull_ten)
         if outcomes is None:
